cambiando types and modify when use this
